Refactor modal structure with the new discord modal format
