+++ package.json
@@ -37,7 +37,7 @@
     "chalk": "^5.4.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
-    "discord.js": "^14.22.1",
+    "discord.js": "^14.23.2",
     "expr-eval": "^2.0.2",
     "glob": "^11.0.1",
     "jsonwebtoken": "^9.0.2",
