+++ src/core/contracts/index.ts
@@ -1,4 +1,4 @@
-import { ActionRowData, APIMessageTopLevelComponent, BitFieldResolvable, APIEmbed, Attachment, AttachmentBuilder, BufferResolvable, MessageMentionOptions, PollData, MessageActionRowComponentBuilder, MessageActionRowComponentData, TopLevelComponentData, JSONEncodable, ActionRowBuilder, ContainerBuilder, FileBuilder, MediaGalleryBuilder, SectionBuilder, SeparatorBuilder, TextDisplayBuilder,  } from 'discord.js';
+import { ActionRowData, APIMessageTopLevelComponent, BitFieldResolvable, APIEmbed, Attachment, AttachmentBuilder, BufferResolvable, MessageMentionOptions, PollData, MessageActionRowComponentBuilder, MessageActionRowComponentData, TopLevelComponentData, JSONEncodable, ActionRowBuilder, ContainerBuilder, FileBuilder, MediaGalleryBuilder, SectionBuilder, SeparatorBuilder, TextDisplayBuilder, TextInputBuilder, StringSelectMenuBuilder, MentionableSelectMenuBuilder, ChannelSelectMenuBuilder, RoleSelectMenuBuilder, UserSelectMenuBuilder,  } from 'discord.js';
 import { Manager, Addon, Button, SelectMenu, Command, ContextMenu, Modal, Logger } from '@itsmybot';
 import { Stream } from 'stream';
 
@@ -9,7 +9,6 @@ export * from './decorators/validator.js';
 export * from './config/baseConfig.js';
 export * from './config/baseConfigSection.js';
 export * from './validators/command.js';
-export * from './validators/message.js';
 export * from './validators/component.js';
 export * from './validators/scripting.js';
 export * from './config/config.js';
@@ -67,3 +66,5 @@ export type ResolvableInteraction =
   | Command<Addon | undefined>
   | ContextMenu<Addon | undefined>
   | Modal<Addon | undefined>;
+
+export type LabelComponentBuilder = TextInputBuilder | StringSelectMenuBuilder | MentionableSelectMenuBuilder | ChannelSelectMenuBuilder | RoleSelectMenuBuilder | UserSelectMenuBuilder
